refactor(api): 合并记忆体系统API端点
- 创建统一的memory.py管理所有记忆体相关操作
- 将persons、organizations、relationships三个文件合并
- 18个端点整合到一个文件，使用子路由组织
- 更新前端页面API调用路径配合新结构
- 清理GraphService中未使用的导入和TODO注释

架构优化：
- 记忆体系统作为统一领域概念
- 减少文件数量：3个文件→1个文件
- 便于维护和扩展
